requirements.txt
tensorflow==2.16.1
matplotlib
numpy
pandas
pillow~=10.3.0
fastapi~=0.111.0
motor~=3.4.0
pymongo~=4.7.3
pydantic~=2.7.4
uvicorn~=0.30.1